+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 * BREAKING: Kafka::FFI::Client#metadata now returns Kafka::Metadata::Cluster
 * BREAKING: Kafka::FFI::Client#group_list now returns array of Kafka::Metadata::Group
-* Update librdkafka to 1.4.2
+* BREAKING: Removed Kafka::LIBRDKAFKA_VERSION and Kafka::LIBRDKAFKA_CHECKSUM
+* Update librdkafka to 1.5.0
 * Fixes Kafka::FFI::Consumer#commit_message
 * Adds kip-511 for client name and version
 * Adds bindings for fnv1a partitioners
+
+### Removal of Kafka::LIBRDKAFKA_VERSION
+
+Prefer to use Kafka::FFI.version to get the current version of librdkafka.
 
 ## 0.5.2 / 2020-01-27